+++ src/library/adash-errors.adb
@@ -27,6 +27,7 @@ package body Adash.Errors is
             | Error_File_Too_Large
             | Error_Output_Too_Large
             | Error_Job_Unknown
+            | Error_Process_Would_Not_Stop
             | Error_Job_Is_Suspended
             | Error_Cancelled
             | Error_Directory_Not_Found
@@ -190,6 +191,7 @@ package body Adash.Errors is
          when Error_Output_Too_Large       =>
             return Adash.Messages.Msg_Output_Too_Large;
          when Error_Job_Unknown            => return Adash.Messages.Msg_Job_Unknown;
+         when Error_Process_Would_Not_Stop => return Adash.Messages.Msg_Process_Would_Not_Stop;
          when Error_Job_Is_Suspended       =>
             return Adash.Messages.Msg_Job_Is_Suspended;
          when Error_Cancelled              => return Adash.Messages.Msg_Execution_Cancelled;